Typos and language errors, inconsistencies in qm.pm
I noticed the typo "stopps" when running `qm stop`, and after tracking it down to this file, since it wasn't too long I figured I might as well do a bit of proof-reading. I can't guarantee I found everything, and you might not agree with all of my suggestions, but here they are anyway. Sorry to be such a pedant.