test: cfg2cmd: don't use QEMU binary version less than 5.0 for tests
Since there are certain checks that depend on the QEMU binary version,
tests with a fixed QEMU binary version make it less likely to catch
issues on current setups, because for those, the QEMU binary version
will always be higher than in the tests.

Set the machine version, because these tests depend on that. The machine
version is what should actually be tested for rather than an old QEMU
binary version.

Signed-off-by: Fiona Ebner <f.ebner@proxmox.com>
Reviewed-by: Daniel Kral <d.kral@proxmox.com>
Tested-by: Daniel Kral <d.kral@proxmox.com>